+++ design-system/progress-indicator/examples/progress-indicator-size-spacing.tsx
@@ -1,5 +1,4 @@
-/* eslint-disable @repo/internal/react/no-class-components */
-import React, { Component } from 'react';
+import React, { FC, useState } from 'react';
 
 import Button from '@atlaskit/button/standard-button';
 import Box from '@atlaskit/ds-explorations/box';
@@ -20,11 +19,7 @@ const SpreadInlineLayout: React.FC<{ children: React.ReactNode }> = ({
 };
 interface ExampleProps {
   selectedIndex: number;
-  values: Array<string>;
-}
-
-interface State {
-  selectedIndex: number;
+  values: string[];
 }
 
 type Sizes = 'small' | 'default' | 'large';
@@ -32,66 +27,51 @@ type Spacing = 'comfortable' | 'cozy' | 'compact';
 
 const sizes: Sizes[] = ['small', 'default', 'large'];
 const spacings: Spacing[] = ['comfortable', 'cozy', 'compact'];
-class Example extends Component<ExampleProps, State> {
-  static defaultProps = {
-    selectedIndex: 0,
-    values: ['first', 'second', 'third'],
+
+const Example: FC<ExampleProps> = ({ values = ['one', 'two', 'three'] }) => {
+  const [selectedIndex, setSelectedIndes] = useState(0);
+
+  const handlePrev = () => {
+    setSelectedIndes((prevState) => prevState - 1);
   };
 
-  state = {
-    selectedIndex: this.props.selectedIndex,
+  const handleNext = () => {
+    setSelectedIndes((prevState) => prevState + 1);
   };
 
-  handlePrev = () => {
-    this.setState((prevState) => ({
-      selectedIndex: prevState.selectedIndex - 1,
-    }));
-  };
-
-  handleNext = () => {
-    this.setState((prevState) => ({
-      selectedIndex: prevState.selectedIndex + 1,
-    }));
-  };
-
-  render() {
-    const { values } = this.props;
-    const { selectedIndex } = this.state;
-
-    return (
-      <Box paddingInline="scale.200" paddingBlock="scale.200" display="block">
-        <SpreadInlineLayout>
-          <Button isDisabled={selectedIndex === 0} onClick={this.handlePrev}>
-            Prev
-          </Button>
-          <Inline gap="scale.300" testId="vr-hook">
-            {sizes.map((size) => (
-              <Stack gap="scale.200">
-                <Text fontWeight="700">{size}</Text>
-                {spacings.map((space) => (
-                  <Stack gap="scale.100">
-                    <Text>{space}</Text>
-                    <ProgressIndicator
-                      selectedIndex={selectedIndex}
-                      values={values}
-                      size={size}
-                      spacing={space}
-                    />
-                  </Stack>
-                ))}
-              </Stack>
-            ))}
-          </Inline>
-          <Button
-            isDisabled={selectedIndex === values.length - 1}
-            onClick={this.handleNext}
-          >
-            Next
-          </Button>
-        </SpreadInlineLayout>
-      </Box>
-    );
-  }
-}
+  return (
+    <Box paddingInline="scale.200" paddingBlock="scale.200" display="block">
+      <SpreadInlineLayout>
+        <Button isDisabled={selectedIndex === 0} onClick={handlePrev}>
+          Prev
+        </Button>
+        <Inline gap="scale.300" testId="vr-hook">
+          {sizes.map((size) => (
+            <Stack key={size} gap="scale.200">
+              <Text fontWeight="700">{size}</Text>
+              {spacings.map((space) => (
+                <Stack key={space} gap="scale.100">
+                  <Text>{space}</Text>
+                  <ProgressIndicator
+                    selectedIndex={selectedIndex}
+                    values={values}
+                    size={size}
+                    spacing={space}
+                  />
+                </Stack>
+              ))}
+            </Stack>
+          ))}
+        </Inline>
+        <Button
+          isDisabled={selectedIndex === values.length - 1}
+          onClick={handleNext}
+        >
+          Next
+        </Button>
+      </SpreadInlineLayout>
+    </Box>
+  );
+};
 
 export default Example;