+++ editor/editor-common/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/editor-common
 
+## 72.0.1
+
+### Patch Changes
+
+- [`4a2a04ee176`](https://bitbucket.org/atlassian/atlassian-frontend/commits/4a2a04ee176) - Internal refactor of panel color logic to support design token babel plugin
+- [`6455cf006b3`](https://bitbucket.org/atlassian/atlassian-frontend/commits/6455cf006b3) - Builds for this package now pass through a tokens babel plugin, removing runtime invocations of the tokens() function and improving performance.
+
 ## 72.0.0
 
 ### Major Changes