+++ elements/util-data-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/util-data-test
 
+## 17.6.7
+
+### Patch Changes
+
+- [`6455cf006b3`](https://bitbucket.org/atlassian/atlassian-frontend/commits/6455cf006b3) - Builds for this package now pass through a tokens babel plugin, removing runtime invocations of the tokens() function and improving performance.
+
 ## 17.6.6
 
 ### Patch Changes