+++ people-and-teams/focused-task-close-account/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/focused-task-close-account
 
+## 2.0.5
+
+### Patch Changes
+
+- [`6455cf006b3`](https://bitbucket.org/atlassian/atlassian-frontend/commits/6455cf006b3) - Builds for this package now pass through a tokens babel plugin, removing runtime invocations of the tokens() function and improving performance.
+
 ## 2.0.4
 
 ### Patch Changes