+++ people-and-teams/profilecard/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/profilecard
 
+## 18.1.2
+
+### Patch Changes
+
+- [`6455cf006b3`](https://bitbucket.org/atlassian/atlassian-frontend/commits/6455cf006b3) - Builds for this package now pass through a tokens babel plugin, removing runtime invocations of the tokens() function and improving performance.
+
 ## 18.1.1
 
 ### Patch Changes